Add Mappers

+++ src/main/java/com/levanxstore/domain/dto/ExpenseDTO.java
@@ -1,20 +1,18 @@
 package com.levanxstore.domain.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class ExpenseDTO {
-
     private Long id;
     private LocalDateTime date;
     private Long supplierId;
     private String supplierName;
     private String description;
     private Integer quantity;
-    private Double unitCost;
-    private Double totalAmount;
+    private BigDecimal unitCost;
+    private BigDecimal totalAmount;
     private String status;
-
-    //Getters and Setters
 
     public Long getId() { return id; }
     public void setId(Long id) { this.id = id; }
@@ -28,11 +26,10 @@ public class ExpenseDTO {
     public void setDescription(String description) { this.description = description; }
     public Integer getQuantity() { return quantity; }
     public void setQuantity(Integer quantity) { this.quantity = quantity; }
-    public Double getUnitCost() { return unitCost; }
-    public void setUnitCost(Double unitCost) { this.unitCost = unitCost; }
-    public Double getTotalAmount() { return totalAmount; }
-    public void setTotalAmount(Double totalAmount) { this.totalAmount = totalAmount; }
+    public BigDecimal getUnitCost() { return unitCost; }
+    public void setUnitCost(BigDecimal unitCost) { this.unitCost = unitCost; }
+    public BigDecimal getTotalAmount() { return totalAmount; }
+    public void setTotalAmount(BigDecimal totalAmount) { this.totalAmount = totalAmount; }
     public String getStatus() { return status; }
     public void setStatus(String status) { this.status = status; }
-    
 }
